requirements.txt
astroid==1.3.8
click==5.1
colorama==0.3.3
cssselect==0.9.1
logilab-common==1.1.0
lxml==3.4.4
pluggy==0.3.1
py==1.4.30
pylint==1.4.4
pyquery==1.2.9
pytest==2.8.2
PyYAML==3.11
requests==2.8.1
six==1.10.0
tox==2.1.1
virtualenv==13.1.2
wheel==0.24.0
